Enforce shutdown trajectory (#1111)
When a shutdown is initiated at the first timepoint of a subproblem
that is linked to the previous subproblem, it is possible to not honour
the specified shutdown ramp rate.

In the worst case, the unit can directly reach zero generation at that
timepoint (in which the shutdown occurs).

This happens because the shutdown ramp constraint is enforced between
the current timepoint and the next timepoint. Instead, this change
enforces the ramp rate between the previous timepoint and the current
timepoint. This involves add two additional linked parameters -
pmin_mw and pmax_mw.

Closes #1111